+++ src/renderer/utils/frameless/index.ts
@@ -1,2 +1,0 @@
-export * from './FramelessWindow';
-export * from './TitleBar';